requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pyyaml>=6.0
ruamel.yaml>=0.17.0
